Improved ColourBox.get_colour
Now custom defined HEX colours are read correctly and added to the dropdow colour list.
Named colours are read and selected in the colour dropdown list. If colour name does not exists Black is selected.

TODO:

add check in core so that only HEX colours are passed to gradient generators